feat: code formatted and fixing 1573 defect
