+++ src/main/java/com/beis/subsidy/ga/schemes/dbpublishingservice/util/AccessManagementConstant.java
@@ -4,7 +4,8 @@ public class AccessManagementConstant {
 	 public static String SM_ACTIVE="Active";
 	 public static String SM_INACTIVE="Inactive";
     public static String BEIS_ADMIN_ROLE="BEIS Administrator";
-    public static String ROLES[]= {"BEIS Administrator","Granting Authority Administrator","Granting Authority Approver"};
+    public static String ROLES[]= {"BEIS Administrator","Granting Authority Administrator","Granting Authority Approver","Granting Authority Encoder"};
+    public static String ADMIN_ROLES[]= {"BEIS Administrator","Granting Authority Administrator"};
     public static String GA_ADMIN_ROLE="Granting Authority Administrator";
     public static String GA_APPROVER_ROLE="Granting Authority Approver";
     public static String GA_ENCODER_ROLE="Granting Authority Encoder";
