+++ src/main/java/com/beis/subsidy/ga/schemes/dbpublishingservice/util/SearchUtils.java
@@ -140,8 +140,8 @@ public class SearchUtils {
 		String userPrincipleStr = userPrinciple.get("userPrinciple").get(0);
 		try {
 			userPrincipleObj = objectMapper.readValue(userPrincipleStr, UserPrinciple.class);
-			if (!AccessManagementConstant.BEIS_ADMIN_ROLE.equals(userPrincipleObj.getRole())) {
-				throw new UnauthorisedAccessException("You are not authorised to "+ entity);
+			if (!!Arrays.asList(AccessManagementConstant.ADMIN_ROLES).contains(userPrincipleObj.getRole())) {
+				throw new UnauthorisedAccessException("You are not authorised to " + entity);
 			}
 		} catch(JsonProcessingException exception){
 			throw new AccessManagementException(HttpStatus.BAD_REQUEST,"JSON parsing Exception");
