+++ src/main/resources/application.properties
@@ -29,7 +29,7 @@ logging.level.org.hibernate.SQL=info
 logging.level.org.hibernate.type=info
 
 
-loggingComponentName= BEIS_Granting_Authority_API
+loggingComponentName= BEIS_SCHEMES_GA_API
 
 graphApiLoginUrl= https://login.microsoftonline.com
 graphApiUrl= https://graph.microsoft.com
